+++ devops/python/build_package.sh
@@ -1,8 +1,7 @@
 #! /bin/sh
-FILE_NAME="$(basename "$0" 2> /dev/null)"
-BASE_DIR="$( cd "$( dirname "$0" )" >/dev/null 2>&1 && pwd )"
+BASE_DIR="$( cd "$( dirname "$0" )" > /dev/null 2>&1 && pwd )"
 
-. $BASE_DIR/setup_variables.sh
+. "$BASE_DIR/setup_variables.sh"
 
 echo "SoftButterfly CI: Packaging for version ${CI_COMMIT_TAG}"