+++ devops/python/setup_pypi_credentials.sh
@@ -1,10 +1,9 @@
 #! /bin/sh
-FILE_NAME="$(basename "$0" 2> /dev/null)"
-BASE_DIR="$( cd "$( dirname "$0" )" >/dev/null 2>&1 && pwd )"
+BASE_DIR="$( cd "$( dirname "$0" )" > /dev/null 2>&1 && pwd )"
 
-. $BASE_DIR/setup_variables.sh
+. "$BASE_DIR/setup_variables.sh"
 
 echo "SoftButterfly CI: Setting up Pypi credentials"
 
 echo "* Setting up credentials"
-poetry config pypi-token.pypi $PYPI_TOKEN
+poetry config pypi-token.pypi "$PYPI_TOKEN"