+++ devops/python/upload_coverage.sh
@@ -1,15 +1,14 @@
 #! /bin/sh
-FILE_NAME="$(basename "$0" 2> /dev/null)"
-BASE_DIR="$( cd "$( dirname "$0" )" >/dev/null 2>&1 && pwd )"
+BASE_DIR="$( cd "$( dirname "$0" )" > /dev/null 2>&1 && pwd )"
 
 . "${BASE_DIR}/setup_variables.sh"
 
-BASEPYTHON=${CI_JOB_NAME##*:}
+BASEPYTHON="${CI_JOB_NAME##*:}"
 
 echo "SoftButterfly CI: Publish coverage report to Codacy for environment ${BASEPYTHON}"
 
 # Send coverage report only on master branch and job is run for python 3.11
-if [ "$CI_COMMIT_BRANCH" == "master" ] && [ "$BASEPYTHON" == "py311" ]; then
+if [ "$CI_COMMIT_BRANCH" = "master" ] && [ "$BASEPYTHON" = "py311" ]; then
     echo "* Publish coverage report to Codacy is enabled"
 
     echo "* Downloading codacy report uploader."