+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add fields `event` and `source` fields to `ErrorsInboxErrorEventInput` input object type.
 * Remove `AgentFeaturesFilter` object type.
 * Remove field `agent_features` from `DocumentationFields` object type.
+* add tests for newrelic_sb_sdk.utils.query.
 
 ## [0.11.0] - 2023-07-26